+++ src/dataclass_extensions/registrable.py
@@ -11,6 +11,7 @@ R = TypeVar("R", bound="Registrable")
 @dataclass
 class Registrable:
     _registry: ClassVar[dict[str, Type[Registrable]]]
+    _default_type: ClassVar[str | None]
 
     type: dataclasses.InitVar[str | None] = dataclasses.field(
         default=None, kw_only=True, repr=False
@@ -27,16 +28,20 @@ class Registrable:
                     f"Available choices are: {list(cls._registry.keys())}"
                 )
             return super().__new__(cls._registry[type])
+        elif cls._default_type is not None and not hasattr(cls, "registered_name"):
+            return super().__new__(cls._registry[cls._default_type])
         else:
             return super().__new__(cls)
 
     def __init_subclass__(cls, **kwargs):
         super().__init_subclass__(**kwargs)
-        if not hasattr(cls, "_choice_registry"):
+        if not hasattr(cls, "_registry"):
             cls._registry = {}
+        if not hasattr(cls, "_default_type"):
+            cls._default_type = None
 
     @classmethod
-    def register(cls, name: str) -> Callable[[Type[R]], Type[R]]:
+    def register(cls, name: str, default: bool = False) -> Callable[[Type[R]], Type[R]]:
         def register_subclass(subclass: Type[R]) -> Type[R]:
             if not issubclass(subclass, cls):
                 raise TypeError(
@@ -46,6 +51,13 @@ class Registrable:
                 raise TypeError(
                     f"class {subclass.__name__} must be a dataclass in order to register it"
                 )
+            if default:
+                if cls._default_type is not None:
+                    raise ValueError(
+                        f"A default implementation for {cls.__name__} has already been registered"
+                    )
+                else:
+                    cls._default_type = name
 
             fields = [
                 (f.name, f.type, f) for f in dataclasses.fields(subclass) if f.name != "type"  # type: ignore
@@ -94,3 +106,9 @@ class Registrable:
     @classmethod
     def get_registered_names(cls) -> list[str]:
         return list(cls._registry.keys())
+
+    @classmethod
+    def get_default(cls: Type[R]) -> Type[R]:
+        if cls._default_type is None:
+            raise ValueError(f"A default implementation of {cls.__name__} has not been registered")
+        return cls.get_registered_class(cls._default_type)