+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `default: bool` option to `Registrable.register()` for defining a default implementation.
+
 ## [v0.1.0](https://github.com/epwalsh/dataclass-extensions/releases/tag/v0.1.0) - 2025-06-10
 
 ## Added